# docs/swagger.yaml
definitions:
...
  dbmodels.CallPlanActions:
    properties:
      action:
        type: string
      action_type:
        type: string
      amount:
        type: number
      call_plan_merchant_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      merchant_action:
        type: string
      name:
        type: string
      note:
        type: string
      product:
        type: string
      reason:
        type: string
      result:
        type: boolean
      status:
        type: string
      updated_at:
        type: string
    type: object
  dbmodels.CallPlanDescriptions:
    properties:
      action_merchant_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      description_id:
        type: integer
      product_merchant_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  dbmodels.ProductMerchants:
    properties:
      created_at:
        type: string
      product_merchant_id:
        type: integer
      product_merchant_name:
        type: string
      updated_at:
        type: string
    type: object
...
  dbmodels.SalesType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.CallPlanActionDetailReq:
    properties:
      call_plan_action_id:
        type: integer
    type: object
...
  models.CallPlanActionSubmitReq:
    properties:
      call_plan_merchant_id:
        type: integer
      merchant_status:
        type: string
      notes:
        type: string
      status:
        type: string
    type: object
...
  models.CallPlanDescriptionMerchantReq:
    properties:
      action_merchant_id:
        type: integer
      product_merchant_id:
        type: integer
    type: object
...
  models.ChangePhotoPinKTPReq:
    properties:
      id_card:
        type: string
      new_pin:
        type: string
      photo:
        type: string
    type: object
  models.ChangePhotoPinKTPRes:
    properties:
      description_code:
        type: string
      response_code:
        type: string
    type: object
  models.CheckAccountIndomarcoRes:
    properties:
      data:
        type: object
      meta:
        $ref: '#/definitions/models.MetaData'
        type: object
    type: object
...
  models.CheckIndomarcoAccountReq:
    properties:
      customer_code:
        type: string
    type: object
...
  models.IndomarcoLoginReq:
    properties:
      deviceId:
        type: string
      deviceToken:
        type: string
      firebase_token:
        type: string
      idNumber:
        type: string
      pin:
        type: string
      role:
        type: string
      salesId:
        type: string
      version_code:
        type: string
    type: object
  models.IndomarcoLoginRes:
    properties:
      description_code:
        type: string
      email:
        type: string
      force_update:
        type: boolean
      functional_id:
        type: string
      phone:
        type: string
      photo:
        type: string
      response_code:
        type: string
      sales_id:
        type: string
      sales_name:
        type: string
      session_token:
        type: string
      sfa_id:
        type: string
      status:
        type: string
    type: object
...
  models.LoginReq:
    properties:
      deviceId:
        type: string
      deviceToken:
        type: string
      firebase_token:
        type: string
      phoneNumber:
        type: string
      pin:
        type: string
      role:
        type: string
      salesId:
        type: string
      version_code:
        type: string
    type: object
  models.LoginRes:
    properties:
      description_code:
        type: string
      email:
        type: string
      force_update:
        type: boolean
      functional_id:
        type: string
      phone:
        type: string
      photo:
        type: string
      response_code:
        type: string
      sales_id:
        type: string
      sales_name:
        type: string
      sales_type:
        items:
          $ref: '#/definitions/dbmodels.SalesType'
        type: array
      session_token:
        type: string
      sfa_id:
        type: string
      status:
        type: string
    type: object
  models.LoginV23Res:
    properties:
      description_code:
        type: string
      email:
        type: string
      force_update:
        type: boolean
      functional_id:
        type: string
      phone:
        type: string
      photo:
        type: string
      response_code:
        type: string
      sales_id:
        type: string
      sales_name:
        type: string
      sales_type_id:
        type: integer
      session_token:
        type: string
      sfa_id:
        type: string
      status:
        type: string
    type: object
...
  models.MetaData:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Ok
        type: string
      status:
        example: true
        type: boolean
    type: object
...
  models.Response:
    properties:
      data:
        type: object
      meta:
        $ref: '#/definitions/models.MetaData'
        type: object
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /v2.2/call_plan/description_merchant:
    post:
      consumes:
      - application/json
      description: CallPlan Description Merchant
      operationId: CallPlan Description Merchant
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.CallPlanDescriptionMerchantReq'
      produces:
      - application/json
      responses:
        "200":
          description: CallPlan Description Merchant Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/dbmodels.CallPlanDescriptions'
              type: object
      summary: CallPlan Description Merchant
      tags:
      - OTTO SFA
  /v2.2/call_plan/product_merchant/list:
    get:
      consumes:
      - application/json
      description: CallPlan Product Merchant List
      operationId: CallPlan Product Merchant List
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: CallPlan Product Merchant List Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dbmodels.ProductMerchants'
                  type: array
              type: object
      summary: CallPlan Product Merchant List
      tags:
      - OTTO SFA
...
  /v2.2/sfa/auth/change_pin_and_ktp:
    post:
      consumes:
      - application/json
      description: Change Pin And Id Card Photo Profil
      operationId: Change Pin And Id Card Photo Profil
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.ChangePhotoPinKTPReq'
      produces:
      - application/json
      responses:
        "200":
          description: Change Pin And Id Card Photo Profil Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.ChangePhotoPinKTPRes'
              type: object
      summary: Change Pin And Id Card Photo Profil
      tags:
      - OTTO SFA
  /v2.2/sfa/auth/login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: Login v2.2
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Login Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginRes'
              type: object
      summary: Login
      tags:
      - OTTO SFA
...
  /v2.2/sfa/call_plan/action/detail:
    post:
      consumes:
      - application/json
      description: CallPlan Action Detail
      operationId: CallPlan Action Detail v2.2
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.CallPlanActionDetailReq'
      produces:
      - application/json
      responses:
        "200":
          description: CallPlan Action Detail Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/dbmodels.CallPlanActions'
              type: object
      summary: CallPlan Action Detail
      tags:
      - OTTO SFA
...
  /v2.2/sfa/call_plan/action/submit:
    post:
      consumes:
      - application/json
      description: CallPlan Action Submit
      operationId: CallPlan Action Submit v2.2
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.CallPlanActionSubmitReq'
      produces:
      - application/json
      responses:
        "200":
          description: CallPlan Action Submit Response EXAMPLE
          schema:
            $ref: '#/definitions/models.Response'
      summary: CallPlan Action Submit
      tags:
      - OTTO SFA
...
  /v2.2/sfa/check_indomarco_account:
    post:
      consumes:
      - application/json
      description: Check Indomarco Account
      operationId: Check Indomarco Account
      parameters:
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.CheckIndomarcoAccountReq'
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Check Indomarco Account Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.CheckAccountIndomarcoRes'
                  type: array
              type: object
      summary: Check Indomarco Account
      tags:
      - OTTO SFA
...
  /v2.2/sfa/indomarco/login:
    post:
      consumes:
      - application/json
      description: Login Indomarco
      operationId: Login Indomarco
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.IndomarcoLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Indomarco Login Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.IndomarcoLoginRes'
              type: object
      summary: Login Indomarco
      tags:
      - OTTO SFA
...
  /v2.3/sfa/auth/login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: Login v2.3
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Login Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginV23Res'
              type: object
      summary: Login
      tags:
      - OTTO SFA
  /v2.3/sfa/call_plan/action/detail:
    post:
      consumes:
      - application/json
      description: CallPlan Action Detail
      operationId: CallPlan Action Detail
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.CallPlanActionDetailReq'
      produces:
      - application/json
      responses:
        "200":
          description: CallPlan Action Detail Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/dbmodels.CallPlanActions'
              type: object
      summary: CallPlan Action Detail
      tags:
      - OTTO SFA
...
  /v2.3/sfa/call_plan/action/submit:
    post:
      consumes:
      - application/json
      description: CallPlan Action Submit
      operationId: CallPlan Action Submit v2.3
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.CallPlanActionSubmitReq'
      produces:
      - application/json
      responses:
        "200":
          description: CallPlan Action Submit Response EXAMPLE
          schema:
            $ref: '#/definitions/models.Response'
      summary: CallPlan Action Submit
      tags:
      - OTTO SFA
...
  /v2.4/sfa/auth/login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: Login v2.4
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Body
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: Login Response EXAMPLE
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.LoginV23Res'
              type: object
      summary: Login
      tags:
      - OTTO SFA
...
swagger: "2.0"